src/backend/requirements.txt
# Django packages
Django>=4.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
drf-spectacular
django-rest-framework-docs

# AI packages
pandas
numpy
scikit-learn
lightgbm
xgboost
joblib
Pillow